requirements.txt
ruamel-yaml~=0.18.14
pillow~=11.3.0
pydantic-settings~=2.10.1
uvicorn~=0.35.0
fastapi~=0.116.1
pydantic~=2.11.7